+++ lib/offline_sync_core.dart
@@ -1,7 +1,22 @@
-export 'src/core/offline_sync_core.dart';
 
+export 'src/core/offline_sync_core.dart';
+export 'src/core/config.dart';
+
+export 'src/cache/cache_entry.dart';
+export 'src/cache/cache_manager.dart';
 export 'src/cache/cache_policy.dart';
 
-export 'src/storage/hive_storage.dart';
+export 'src/sync/sync_manager.dart';
+export 'src/sync/sync_queue.dart';
+export 'src/sync/sync_task.dart';
+export 'src/sync/sync_status.dart';
 
+export 'src/storage/storage_adapter.dart';
+export 'src/storage/hive_storage.dart';
 export 'src/storage/sqlite_storage.dart';
+
+export 'src/inspector/inspector_controller.dart';
+export 'src/inspector/inspector_screen.dart';
+
+export 'src/utils/logger.dart';
+export 'src/utils/extensions.dart';